pyproject.toml
[project]
name = "hitz11"
version = "0.1.0"
description = "Daily etymology publisher for Telegram channels"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "httpx>=0.27.2",
  "lxml>=5.3.0",
  "pydantic>=2.9.2",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "respx>=0.21.1",
  "ruff>=0.7.0",
]

[project.scripts]
hitz11 = "hitz11.main:main"

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hitz11"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
